[refactor] RacingGame 클래스를 도메인 클래스로 한번더 분리
