+++ src/main/java/project/game/controller/RacingGameController.java
@@ -1,12 +1,11 @@
 package project.game.controller;
 
 import project.game.StringUtils;
-import project.game.domain.Car;
 import project.game.domain.GameType;
-import project.game.domain.MoveRule;
+import project.game.domain.RacingGame;
 import project.game.view.View;
 
-import java.util.*;
+import java.util.List;
 
 import static project.game.TextConstant.*;
 
@@ -15,28 +14,25 @@ public class RacingGameController {
     private static final int INVALID_ROUND_COUNT = -1;
 
     private View mView;
-    private List<Car> mCars;
+    private RacingGame mRacingGame;
 
     public RacingGameController(View view) {
         this.mView = view;
-        this.mCars = new ArrayList<>();
+        this.mRacingGame = new RacingGame();
     }
 
+    // start 를 통해 시작할경우 동적으로 게임 타입을 바꿀 수 있고 시작할 수 있는 시점을 컨트롤할 수 있다.
+    // 이 방식에 단점은 무엇일까? 생성자 초기화 vs set 초기화
     public void start(GameType gameType) {
         mView.drawText(RACING_GAME_START_TEXT);
 
         initializeCars(gameType);
         startRacing();
-        endRacing();
     }
 
     private void initializeCars(GameType gameType) {
         List<String> carNames = getCarNames();
-        MoveRule moveRule = gameType.getMoveRule();
-
-        for (String each : carNames) {
-            mCars.add(new Car(each, moveRule));
-        }
+        mRacingGame.setNewGame(gameType, carNames);
     }
 
     private List<String> getCarNames() {
@@ -53,17 +49,10 @@ public class RacingGameController {
         }
 
         for (int i = 0; i < roundCount; i++) {
-            moveCars();
-            showCurrentCarPositions();
+            showCurrentCarPositions(mRacingGame.moveCars());
         }
 
         showWinners();
-    }
-
-    private void moveCars() {
-        for (Car car : mCars) {
-            car.moveIfPossible();
-        }
     }
 
     private int getRoundCount() {
@@ -75,42 +64,20 @@ public class RacingGameController {
         }
     }
 
-    private void endRacing() {
-        mCars.clear();
-    }
-
     private String readInput(String question) {
         mView.drawText(question);
         return mView.readInput();
     }
 
     private void showWinners() {
-        String winnersText = String.format(RACING_GAME_WINNERS, String.join(COMMA_DELIMITER, findWinnerNames()));
-        mView.drawText(winnersText);
+        mView.drawText(mRacingGame.getWinners());
     }
 
-    private void showCurrentCarPositions() {
+    private void showCurrentCarPositions(List<String> currentPositions) {
         mView.drawNewLine();
 
-        for (Car each : mCars) {
-            mView.drawText(each.getCurrentPositionText());
-        }
-    }
-
-    private List<String> findWinnerNames() {
-        List<String> winnerNames = new ArrayList<>();
-        int maxPosition = Collections.max(mCars, Comparator.comparingInt(Car::getMovePosition)).getMovePosition();
-
-        for (Car each : mCars) {
-            addWinnerNameIfPossible(winnerNames, each, maxPosition);
-        }
-
-        return winnerNames;
-    }
-
-    private void addWinnerNameIfPossible(List<String> winnerNames, Car car, int maxPosition) {
-        if (car.hasEqualPositionTo(maxPosition)) {
-            winnerNames.add(car.getName());
+        for (String each : currentPositions) {
+            mView.drawText(each);
         }
     }
 }
